Extract bundle tests
